gcc/config/pdp1: return `PDP1_IO` in `pdp1_function_value`
Previously, this port had a weird register windowing system where the
function would try putting its return value to `PDP1_ACC`, and then later
the function epilogue would move it to `PDP1_R1`.

This was needed because GCC assumes that it can easily (in one move) do
a `movm` from the register holding the return value to the register
where the caller expects to see it.

The only free register which we can use for that on PDP-1 is the
`PDP1_IO` register, so let's use it for that.

Signed-off-by: Filip Kokosiński <filip.kokosinski@gmail.com>